Sort out cyclic dependencies
